// types/common/response.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.8.3
//   protoc               v6.33.5
// source: common/response.proto

/* eslint-disable */

export const protobufPackage = "common.basic.response";

export interface SuccessResponse {
  message: string;
}

export const COMMON_BASIC_RESPONSE_PACKAGE_NAME = "common.basic.response";